resolve comments; fix imports; extract to wallet; extract to background/utils
