+++ src/background/utils.ts
@@ -8,6 +8,27 @@ import type { Browser, Runtime } from 'webextension-polyfill';
 import { DEFAULT_SCALE, EXCHANGE_RATES_URL } from './config';
 import { INTERNAL_PAGE_URL_PROTOCOLS, NEW_TAB_PAGES } from './constants';
 import { notNullOrUndef } from '@/shared/helpers';
+import { OPEN_PAYMENTS_REDIRECT_URL } from '@/shared/defines';
+
+export const enum GrantResult {
+  GRANT_SUCCESS = 'grant_success',
+  GRANT_ERROR = 'grant_error',
+  KEY_ADD_SUCCESS = 'key_add_success',
+  KEY_ADD_ERROR = 'key_add_error',
+}
+
+export const enum InteractionIntent {
+  CONNECT = 'connect',
+  RECONNECT = 'reconnect',
+  FUNDS = 'funds',
+  UPDATE_BUDGET = 'update_budget',
+}
+
+export const enum ErrorCode {
+  CONTINUATION_FAILED = 'continuation_failed',
+  HASH_FAILED = 'hash_failed',
+  KEY_ADD_FAILED = 'key_add_failed',
+}
 
 export const getCurrentActiveTab = async (browser: Browser) => {
   const window = await browser.windows.getLastFocused();
@@ -84,6 +105,31 @@ export const getTabId = (sender: Runtime.MessageSender): number => {
 
 export const getTab = (sender: Runtime.MessageSender): Tab => {
   return notNullOrUndef(notNullOrUndef(sender.tab, 'sender.tab'), 'tab') as Tab;
+};
+
+export const redirectToWelcomeScreen = async (
+  browser: Browser,
+  tabId: number,
+  result: GrantResult,
+  intent: InteractionIntent,
+  errorCode?: ErrorCode,
+): Promise<void> => {
+  const url = new URL(OPEN_PAYMENTS_REDIRECT_URL);
+  url.searchParams.set('result', result);
+  url.searchParams.set('intent', intent);
+  if (errorCode) url.searchParams.set('errorCode', errorCode);
+
+  await browser.tabs.update(tabId, {
+    url: url.toString(),
+  });
+};
+
+export const ensureTabExists = async (browser: Browser): Promise<number> => {
+  const tab = await browser.tabs.create({});
+  if (!tab.id) {
+    throw new Error('Could not create tab');
+  }
+  return tab.id;
 };
 
 export const reuseOrCreateTab = async (
